disease_assessment/requirements/base.txt
Django==2.2.7
django-environ==0.4.5
django-filter==2.2.0
djangorestframework==3.10.3
psycopg2==2.8.4
pytz==2019.3
sqlparse==0.3.0
pandas==1.1.4
tqdm==4.52.0